fix(neofetch): list all Intel GPUs as detected
Now that Intel is selling its own dedicated GPUs, it is increasingly
common to see laptops and desktop devices with multiple Intel GPUs. This
patch is intended to fix GPU detection on a laptop with both an integrated
UHD Graphics GPU and an Arc 380 (DG2).

Note that this may potentially cause regressions, but the original
upstream did not record the original issue, so I'm inclined to write it
off as an outdated fix:

  commit ee815f9c6651423ae93f8ead237ebc99e2373c06
  Author: Dylan Araps <dylan.araps@gmail.com>
  Date:   Mon Apr 9 09:51:23 2018 +1000

      gpu: Fixed duplicate intel bug.